requirements.txt
asgiref==3.3.1
Django==3.1.7
django-debug-toolbar==3.2
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
